==> scripts/1_data_filtering.R <==
# Arc-well - filtering datasets
#---Note: run mapping and segmention scripts before running this-----#######
#---mapping and segmentation scripts in https://github.com/navinlabcode/CNV_pipeline
#---This script needs input files from ./final_result/ folder and ./metrics folder generated by above CNA_pipeline: 
# uber.*.bin.txt 
# uber.*.ratio.txt 
# uber.*.seg.txt 
# all_stat_metrics.txt

#####-------loading packages-----------################
library(tidyverse)
library(stringr)
library(dplyr)
library(tidyr)
library(cowplot)
library(useful)
library(tibble)
library(copykit)
library(SummarizedExperiment)
options(max.print = 200)

setwd("/volumes/USR2/wangkl/wafergen/DNA/ffpe_dcis/github_upload")
load("./pre_load_data/pre_load_data.rda")
source("./scripts/0_arcwell_functions.R")
bin_coords <- read.csv("./pre_load_data/bin_coords.csv")
######---- Arc-well-mda231 -----########
#-folder contains ./final_result and ./metrics
met_path <- c("./map_seg_output/arc_well_mda231")
#-convert cell names in seg file into original names---
system(paste0("sed -i '1s/\\./-/g' ", met_path, "/final_result/uber.*.seg.txt"))

arc_well_mda231_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

arc_well_mda231 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
arc_well_mda231 <- filterCells(arc_well_mda231, resolution = 0.8)

arc_well_mda231_metadata <- as.data.frame(colData(arc_well_mda231))
arc_well_mda231_metadata_metrics <- left_join(arc_well_mda231_metadata, arc_well_mda231_metrics)
write.table(arc_well_mda231_metadata_metrics,"./metrics/arc_well_mda231_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

arc_well_mda231 <- arc_well_mda231[,SummarizedExperiment::colData(arc_well_mda231)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(arc_well_mda231,'bin_counts')), 
            "./metrics/arc_well_mda231_filtered_bincounts_newnormal.txt",row.names = F)

# 10X genomics Tn17 ----
met_path <- c("./map_seg_output/tenx")
tenx_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 
tenx <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)

tenx <- filterCells(tenx, resolution = 0.8)
tenx_metadata <- as.data.frame(colData(tenx))
tenx_metadata_metrics <- left_join(tenx_metadata, tenx_metrics)
write.table(tenx_metadata_metrics,"./metrics/tenx_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

tenx <- tenx[,SummarizedExperiment::colData(tenx)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(tenx,'bin_counts')), "./metrics/tenx_filtered_bincounts_newnormal.txt",row.names = F)


# DOP-PCR-merged----
met_path <- c("./map_seg_output/dop_merge")
dop_merge_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

dop_merge <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
dop_merge <- filterCells(dop_merge, resolution = 0.8)

dop_merge_metadata <- as.data.frame(colData(dop_merge))
# joining with metrics
dop_merge_metadata_metrics <- left_join(dop_merge_metadata, dop_merge_metrics)
write.table(dop_merge_metadata_metrics,"./metrics/dop_merge_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

dop_merge <- dop_merge[,SummarizedExperiment::colData(dop_merge)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(dop_merge,'bin_counts')), "./metrics/dop_merge_filtered_bincounts_newnormal.txt",row.names = F)

# DLP-merged----
met_path <- c("./map_seg_output/dlp_merge")
dlp_merge_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

dlp_merge <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
dlp_merge <- filterCells(dlp_merge, resolution = 0.8)

dlp_merge_metadata <- as.data.frame(colData(dlp_merge))

# joining with metrics
dlp_merge_metadata_metrics <- left_join(dlp_merge_metadata, dlp_merge_metrics)
write.table(dlp_merge_metadata_metrics,"./metrics/dlp_merge_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

dlp_merge <- dlp_merge[,SummarizedExperiment::colData(dlp_merge)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(dlp_merge,'bin_counts')), "./metrics/dlp_merge_filtered_bincounts_newnormal.txt",row.names = F)

# DLP+ ----
met_path <- c("./map_seg_output/dlp_plus")

dlp_plus_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

dlp_plus <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
dlp_plus <- filterCells(dlp_plus, resolution = 0.8)

dlp_plus_metadata <- as.data.frame(colData(dlp_plus))

# joining with metrics
dlp_plus_metadata_metrics <- left_join(dlp_plus_metadata, dlp_plus_metrics)
write.table(dlp_plus_metadata_metrics,"./metrics/dlp_plus_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

dlp_plus <- dlp_plus[,SummarizedExperiment::colData(dlp_plus)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(dlp_plus,'bin_counts')), "./metrics/dlp_plus_filtered_bincounts_newnormal.txt",row.names = F)


# ACT MDA231----
met_path <- c("./map_seg_output/act_mda231")
act_mda231_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

act_mda231 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
act_mda231 <- filterCells(act_mda231, resolution = 0.8)

act_mda231_metadata <- as.data.frame(colData(act_mda231))
act_mda231_metadata_metrics <- left_join(act_mda231_metadata, act_mda231_metrics)
write.table(act_mda231_metadata_metrics,"./metrics/act_mda231_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

act_mda231 <- act_mda231[,SummarizedExperiment::colData(act_mda231)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(act_mda231,'bin_counts')), "./metrics/act_mda231_filtered_bincounts_newnormal.txt",row.names = F)

# 315A_fresh-------
# reading metrics
met_path <- c("./map_seg_output/fresh_315a")
fresh_315a_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

fresh_315a <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
fresh_315a <- filterCells(fresh_315a, resolution = 0.8)

fresh_315a_metadata <- as.data.frame(colData(fresh_315a))

# joining with metrics
fresh_315a_metadata_metrics <- left_join(fresh_315a_metadata, fresh_315a_metrics)
write.table(fresh_315a_metadata_metrics,"./metrics/fresh_315a_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

fresh_315a <- fresh_315a[,SummarizedExperiment::colData(fresh_315a)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(fresh_315a,'bin_counts')), "./metrics/fresh_315a_filtered_bincounts_newnormal.txt",row.names = F)

# 315A_formalin-------
met_path <- c("./map_seg_output/formalin_315a")
formalin_315a_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

formalin_315a <- readVarbinCNA(met_path, clean_names = FALSE, remove_Y = T)
formalin_315a <- filterCells(formalin_315a, resolution = 0.8)

formalin_315a_metadata <- as.data.frame(colData(formalin_315a))
# joining with metrics
formalin_315a_metadata_metrics <- left_join(formalin_315a_metadata, formalin_315a_metrics)
write.table(formalin_315a_metadata_metrics,"./metrics/formalin_315a_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

formalin_315a <- formalin_315a[,SummarizedExperiment::colData(formalin_315a)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(formalin_315a,'bin_counts')), "./metrics/formalin_315a_filtered_bincounts_newnormal.txt",row.names = F)

# MDA231_fresh-------
met_path <- c("./map_seg_output/fresh_mda231")
fresh_mda231_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

fresh_mda231 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
fresh_mda231 <- filterCells(fresh_mda231, resolution = 0.8)

fresh_mda231_metadata <- as.data.frame(colData(fresh_mda231))

# joining with metrics
fresh_mda231_metadata_metrics <- left_join(fresh_mda231_metadata, fresh_mda231_metrics)
write.table(fresh_mda231_metadata_metrics,"./metrics/fresh_mda231_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

fresh_mda231 <- fresh_mda231[,SummarizedExperiment::colData(fresh_mda231)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(fresh_mda231,'bin_counts')), "./metrics/fresh_mda231_filtered_bincounts_newnormal.txt",row.names = F)

# MDA231_formalin-------
met_path <- c("./map_seg_output/formalin_mda231")
formalin_mda231_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

formalin_mda231 <- readVarbinCNA(met_path, clean_names = FALSE, remove_Y = T)
formalin_mda231 <- filterCells(formalin_mda231, resolution = 0.8)

formalin_mda231_metadata <- as.data.frame(colData(formalin_mda231))
# joining with metrics
formalin_mda231_metadata_metrics <- left_join(formalin_mda231_metadata, formalin_mda231_metrics)
write.table(formalin_mda231_metadata_metrics,"./metrics/formalin_mda231_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

formalin_mda231 <- formalin_mda231[,SummarizedExperiment::colData(formalin_mda231)$filtered == "kept"]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(formalin_mda231,'bin_counts')), "./metrics/formalin_mda231_filtered_bincounts_newnormal.txt",row.names = F)

# bcis28(DCIS)_frozen_QC ----
met_path <- c("./map_seg_output/bcis28_frozen_qc")
bcis28_frozen_qc_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

bcis28_frozen_qc <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
bcis28_frozen_qc <- filterCells(bcis28_frozen_qc, resolution = 0.8)
bcis28_frozen_qc <- findNormalCells(bcis28_frozen_qc)

bcis28_frozen_qc_metadata <- as.data.frame(colData(bcis28_frozen_qc))
# joining with metrics
bcis28_frozen_qc_metadata_metrics <- left_join(bcis28_frozen_qc_metadata, bcis28_frozen_qc_metrics)
write.table(bcis28_frozen_qc_metadata_metrics,"./metrics/bcis28_frozen_qc_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

bcis28_frozen_qc <- bcis28_frozen_qc[,SummarizedExperiment::colData(bcis28_frozen_qc)$filtered == "kept"]
bcis28_frozen_qc@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
bcis28_frozen_qc <- bcis28_frozen_qc[,SummarizedExperiment::colData(bcis28_frozen_qc)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(bcis28_frozen_qc,'bin_counts')), 
            "./metrics/bcis28_frozen_qc_filtered_bincounts_newnormal.txt",row.names = F)

# bcis28(DCIS)_ffpe_QC ----
met_path <- c("./map_seg_output/bcis28_ffpe_qc")
bcis28_ffpe_qc_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>% 
  dplyr::rename(sample = "Sample.Name") %>% mutate(dups_percentage = DupsRemoved/TotalReads) 

bcis28_ffpe_qc <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
bcis28_ffpe_qc <- filterCells(bcis28_ffpe_qc, resolution = 0.8)
bcis28_ffpe_qc <- findAneuploid(bcis28_ffpe_qc)

bcis28_ffpe_qc_metadata <- as.data.frame(colData(bcis28_ffpe_qc))
# joining with metrics
bcis28_ffpe_qc_metadata_metrics <- left_join(bcis28_ffpe_qc_metadata, bcis28_ffpe_qc_metrics)
write.table(bcis28_ffpe_qc_metadata_metrics,"./metrics/bcis28_ffpe_qc_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

bcis28_ffpe_qc <- bcis28_ffpe_qc[,SummarizedExperiment::colData(bcis28_ffpe_qc)$filtered == "kept"]
bcis28_ffpe_qc@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
bcis28_ffpe_qc <- bcis28_ffpe_qc[,SummarizedExperiment::colData(bcis28_ffpe_qc)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(bcis28_ffpe_qc,'bin_counts')), 
            "./metrics/bcis28_ffpe_qc_filtered_bincounts_newnormal.txt",row.names = F)


# sc97_30 (lung30, lung-p1) ----
met_path <- c("./map_seg_output/lung30_p1")
sc97_30_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                              sep = "\t",
                              header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

sc97_30 <- readVarbinCNA(met_path, 
                         remove_Y = TRUE,
                         clean_names = FALSE)

sc97_30 <- filterCells(sc97_30, resolution = 0.8)
sc97_30 <- findNormalCells(sc97_30)
sc97_30_metadata <- as.data.frame(colData(sc97_30))

# joining with metrics
sc97_30_metadata_metrics <- left_join(sc97_30_metadata, sc97_30_metrics)
write.table(sc97_30_metadata_metrics,"./metrics/lung30_p1_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

sc97_30 <- sc97_30[,SummarizedExperiment::colData(sc97_30)$filtered == "kept"]
sc97_30@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
sc97_30 <- sc97_30[,SummarizedExperiment::colData(sc97_30)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(sc97_30,'bin_counts')), "./metrics/lung30_p1_filtered_bincounts_newnormal.txt",row.names = F)


# sc97_31 (lung31, lung-p2)----
met_path <- c("./map_seg_output/lung31_p2")
sc97_31_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                              sep = "\t",
                              header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

sc97_31 <- readVarbinCNA(met_path, 
                         remove_Y = TRUE,
                         clean_names = FALSE)

sc97_31 <- filterCells(sc97_31, resolution = 0.8)
sc97_31 <- findNormalCells(sc97_31)

sc97_31_metadata <- as.data.frame(colData(sc97_31))
# joining with metrics
sc97_31_metadata_metrics <- left_join(sc97_31_metadata, sc97_31_metrics)
write.table(sc97_31_metadata_metrics,"./metrics/lung31_p2_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

sc97_31 <- sc97_31[,SummarizedExperiment::colData(sc97_31)$filtered == "kept"]
sc97_31@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
sc97_31 <- sc97_31[,SummarizedExperiment::colData(sc97_31)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(sc97_31,'bin_counts')), "./metrics/lung31_p2_filtered_bincounts_newnormal.txt",row.names = F)


# ax4_bl (prostate-p1)----
met_path <- c("./map_seg_output/prostate_p1_ax4bl")
ax4_bl_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                             sep = "\t",
                             header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

ax4_bl <- readVarbinCNA(met_path, 
                        remove_Y = TRUE,
                        clean_names = FALSE)

#---remove cells from the other sample (ax62)
ax4_bl <- ax4_bl[,stringr::str_detect(ax4_bl@colData$sample, "AX4")]
ax4_bl <- filterCells(ax4_bl, resolution = 0.8)
ax4_bl <- findNormalCells(ax4_bl)

ax4_bl_metadata <- as.data.frame(colData(ax4_bl))

# joining with metrics
ax4_bl_metadata_metrics <- left_join(ax4_bl_metadata, ax4_bl_metrics)
write.table(ax4_bl_metadata_metrics,"./metrics/prostate_p1_ax4bl_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

ax4_bl <- ax4_bl[,SummarizedExperiment::colData(ax4_bl)$filtered == "kept"]
ax4_bl@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
ax4_bl <- ax4_bl[,SummarizedExperiment::colData(ax4_bl)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2, assay(ax4_bl,'bin_counts')), "./metrics/prostate_p1_ax4bl_filtered_bincounts_newnormal.txt",row.names = F)


# pcf_169 (prostate-p2)----
met_path <- c("./map_seg_output/prostate_p2_pcf169")
pcf_169_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                              sep = "\t",
                              header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

pcf_169 <- readVarbinCNA(met_path, 
                         remove_Y = TRUE,
                         clean_names = FALSE)
pcf_169 <- filterCells(pcf_169, resolution = 0.8)
pcf_169 <- findNormalCells(pcf_169)

pcf_169_metadata <- as.data.frame(colData(pcf_169))
# joining with metrics
pcf_169_metadata_metrics <- left_join(pcf_169_metadata, pcf_169_metrics)
write.table(pcf_169_metadata_metrics,"./metrics/prostate_p2_pcf169_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

pcf_169 <- pcf_169[,SummarizedExperiment::colData(pcf_169)$filtered == "kept"]
pcf_169@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
pcf_169 <- pcf_169[,SummarizedExperiment::colData(pcf_169)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(pcf_169,'bin_counts')), "./metrics/prostate_p2_pcf169_filtered_bincounts_newnormal.txt",row.names = F)


# DUK249 (P1)----
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Tue Jan 12 15:55:40 2021
met_path <- c("./map_seg_output/duke249_p1")
duke249 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
duke249 <- filterCells(duke249, resolution = .8)
duke249 <- findNormalCells(duke249)

# reading metrics
duke249_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>%
  janitor::clean_names() %>%
  dplyr::rename(sample = "sample_name") %>%
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

colData(duke249)$timepoint <- ifelse(str_detect(colData(duke249)$sample, "DUKE_27"), "recurrence","primary")

duke249_metadata <- as.data.frame(colData(duke249))
# joining with metrics
duke249_metadata_metrics <- left_join(duke249_metadata, duke249_metrics)
write.table(duke249_metadata_metrics, "./metrics/duke249_p1_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

duke249 <- duke249[,SummarizedExperiment::colData(duke249)$filtered == "kept"]
duke249@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
duke249 <- duke249[,SummarizedExperiment::colData(duke249)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(duke249,'bin_counts')), "./metrics/duke249_p1_filtered_bincounts_newnormal.txt",row.names = F)


# NKI17 (P2)----
met_path <- c("./map_seg_output/nki17_p2")
nki17 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki17 <- filterCells(nki17, resolution = .8)
nki17 <- findNormalCells(nki17)

# reading metrics
nki17_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>%
  janitor::clean_names() %>%
  dplyr::rename(sample = "sample_name") %>%
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

colData(nki17)$timepoint <- ifelse(str_detect(colData(nki17)$sample, "NKI17P"), "primary", "recurrence")

nki17_metadata <- as.data.frame(colData(nki17))
# joining with metrics
nki17_metadata_metrics <- left_join(nki17_metadata, nki17_metrics)
write.table(nki17_metadata_metrics,"./metrics/nki17_p2_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

nki17 <- nki17[,SummarizedExperiment::colData(nki17)$filtered == "kept"]
nki17@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki17 <- nki17[,SummarizedExperiment::colData(nki17)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki17,'bin_counts')), "./metrics/nki17_p2_filtered_bincounts_newnormal.txt",row.names = F)

# DUK248 (P3) ----
met_path <- c("./map_seg_output/duke248_p3")
preduk248 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
preduk248 <- filterCells(preduk248, resolution = .8)
preduk248 <- findNormalCells(preduk248)

colData(preduk248)$timepoint <- ifelse(str_detect(colData(preduk248)$sample, "DUKE_24"), "primary", "recurrence")

# reading metrics
preduk248_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"), sep = "\t", header = T) %>%
  janitor::clean_names() %>%
  dplyr::rename(sample = "sample_name") %>%
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

preduk248_metadata <- as.data.frame(colData(preduk248))
# joining with metrics
preduk248_metadata_metrics <- left_join(preduk248_metadata, preduk248_metrics)
write.table(preduk248_metadata_metrics, "./metrics/duke248_p3_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

preduk248 <- preduk248[,SummarizedExperiment::colData(preduk248)$filtered == "kept"]
preduk248@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
preduk248 <- preduk248[,SummarizedExperiment::colData(preduk248)$is_normal == FALSE]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(preduk248,'bin_counts')), "./metrics/duke248_p3_filtered_bincounts_newnormal.txt",row.names = F)

# DUK254 (P4) ----
met_path <- c("./map_seg_output/duke254_p4")
duke254 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
duke254 <- filterCells(duke254, resolution = 0.8)
duke254 <- findNormalCells(duke254)
colData(duke254)$timepoint <- ifelse(str_detect(colData(duke254)$sample, "DUKE_30"), "primary", "recurrence")

duke254_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                              sep = "\t",
                              header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))


duke254_metadata <- as.data.frame(colData(duke254))
# joining with metrics
duke254_metadata_metrics <- left_join(duke254_metadata, duke254_metrics)
write.table(duke254_metadata_metrics,"./metrics/duke254_p4_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

duke254 <- duke254[,SummarizedExperiment::colData(duke254)$filtered == "kept"]
duke254@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
duke254 <- duke254[,SummarizedExperiment::colData(duke254)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(duke254,'bin_counts')), "./metrics/duke254_p4_filtered_bincounts_newnormal.txt",row.names = F)


# NKI23 (P5) ----
met_path <- c("./map_seg_output/nki23_p5")
nki23_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

nki23 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki23 <- filterCells(nki23, resolution = 0.8)
nki23 <- findNormalCells(nki23)

colData(nki23)$timepoint <- ifelse(str_detect(colData(nki23)$sample, "RE"), "recurrence", "primary")
nki23_metadata <- as.data.frame(colData(nki23))

# joining with metrics
nki23_metadata_metrics <- left_join(nki23_metadata, nki23_metrics)
write.table(nki23_metadata_metrics, "./metrics/nki23_p5_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)
nki23 <- nki23[,SummarizedExperiment::colData(nki23)$filtered == "kept"]
nki23@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki23 <- nki23[,SummarizedExperiment::colData(nki23)$is_normal == FALSE]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki23,'bin_counts')), "./metrics/nki23_p5_filtered_bincounts_newnormal.txt", row.names = F)

# NKI26 (P6) ----
met_path <- c("./map_seg_output/nki26_p6")
nki26 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki26 <- filterCells(nki26, resolution = .8)
nki26 <- findNormalCells(nki26)

# reading metrics
nki26_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                                 sep = "\t",
                                 header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace(sample, "-", "\\."))

colData(nki26)$timepoint <- ifelse(str_detect(colData(nki26)$sample, "RE"), "recurrence", "primary")
nki26_metadata <- as.data.frame(colData(nki26))
# joining with metrics
nki26_metadata_metrics <- left_join(nki26_metadata, nki26_metrics)
write.table(nki26_metadata_metrics, "./metrics/nki26_p6_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki26 <- nki26[,SummarizedExperiment::colData(nki26)$filtered == "kept"]
nki26@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki26 <- nki26[,SummarizedExperiment::colData(nki26)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki26, 'bin_counts')), "./metrics/nki26_p6_filtered_bincounts_newnormal.txt", row.names = F)

# NKI28 (P7) -----
met_path <- c("./map_seg_output/nki28_p7")
nki28 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki28 <- filterCells(nki28, resolution = .8)
nki28 <- findNormalCells(nki28)
# reading metrics
nki28_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace(sample, "-", "\\.")) 

colData(nki28)$timepoint <- ifelse(str_detect(colData(nki28)$sample, "RE"), "recurrence", "primary")
nki28_metadata <- as.data.frame(colData(nki28))

# joining with metrics
nki28_metadata_metrics <- left_join(nki28_metadata, nki28_metrics)
write.table(nki28_metadata_metrics, "./metrics/nki28_p7_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki28 <- nki28[,SummarizedExperiment::colData(nki28)$filtered == "kept"]
nki28@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki28 <- nki28[,SummarizedExperiment::colData(nki28)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2, assay(nki28, 'bin_counts')), "./metrics/nki28_p7_filtered_bincounts_newnormal.txt", row.names = F)

# NKI19 (P8) ----
met_path <- c("./map_seg_output/nki19_p8")
nki19 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki19 <- filterCells(nki19, resolution = .8)
nki19 <- findNormalCells(nki19)

colData(nki19)$timepoint <- ifelse(str_detect(colData(nki19)$sample, "RE"), "recurrence", "primary")
# reading metrics
nki19_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace(sample, "-", "\\."))


nki19_metadata <- as.data.frame(colData(nki19))
# joining with metrics
nki19_metadata_metrics <- left_join(nki19_metadata, nki19_metrics)
write.table(nki19_metadata_metrics, "./metrics/nki19_p8_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki19 <- nki19[,SummarizedExperiment::colData(nki19)$filtered == "kept"]
nki19@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki19 <- nki19[,SummarizedExperiment::colData(nki19)$is_normal == FALSE]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki19, 'bin_counts')), "./metrics/nki19_p8_filtered_bincounts_newnormal.txt", row.names = F)

# NKI22 (P9) ----
met_path <- c("./map_seg_output/nki22_p9")
nki22_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                                         sep = "\t",
                                         header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

nki22 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki22 <- filterCells(nki22, resolution = 0.8)
nki22 <- findNormalCells(nki22)
colData(nki22)$timepoint <- ifelse(str_detect(colData(nki22)$sample, "RE"), "recurrence", "primary")

nki22_metadata <- as.data.frame(colData(nki22))
# joining with metrics
nki22_metadata_metrics <- left_join(nki22_metadata, nki22_metrics)
write.table(nki22_metadata_metrics,"./metrics/nki22_p9_metadata.metrics_newnormal.txt", sep = "\t",row.names = F)

nki22 <- nki22[,SummarizedExperiment::colData(nki22)$filtered == "kept"]
nki22@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki22 <- nki22[,SummarizedExperiment::colData(nki22)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki22, 'bin_counts')), "./metrics/nki22_p9_filtered_bincounts_newnormal.txt",row.names = F)

# NKI15 (P10)  -----
met_path <- c("./map_seg_output/nki15_p10")
nki15 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki15 <- filterCells(nki15, resolution = .8)
nki15 <- findAneuploid(nki15)

# reading metrics
nki15_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

colData(nki15)$timepoint <- ifelse(str_detect(colData(nki15)$sample, "RE"), "recurrence", "primary")
nki15_metadata <- as.data.frame(colData(nki15))
# joining with metrics
nki15_metadata_metrics <- left_join(nki15_metadata, nki15_metrics)
write.table(nki15_metadata_metrics, "./metrics/nki15_p10_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki15 <- nki15[,SummarizedExperiment::colData(nki15)$filtered == "kept"]
table(nki15@colData %>% as.data.frame() %>% pull(is_normal), nki15@colData %>% as.data.frame() %>% pull(timepoint))
nki15 <- nki15[,SummarizedExperiment::colData(nki15)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki15, 'bin_counts')), "./metrics/nki15_p10_filtered_bincounts_newnormal.txt", row.names = F)

# NKI31 (P11) -----
met_path <- c("./map_seg_output/nki31_p11")
nki31 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki31 <- filterCells(nki31, resolution = .8)
nki31 <- findNormalCells(nki31)

# reading metrics
nki31_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace(sample, "-", "\\."))

colData(nki31)$timepoint <- ifelse(str_detect(colData(nki31)$sample, "RE"), "recurrence", "primary")
nki31_metadata <- as.data.frame(colData(nki31))
# joining with metrics
nki31_metadata_metrics <- left_join(nki31_metadata, nki31_metrics)
write.table(nki31_metadata_metrics, "./metrics/nki31_p11_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki31 <- nki31[,SummarizedExperiment::colData(nki31)$filtered == "kept"]
nki31@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki31 <- nki31[,SummarizedExperiment::colData(nki31)$is_normal == FALSE]

bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2,assay(nki31, 'bin_counts')), "./metrics/nki31_p11_filtered_bincounts_newnormal.txt", row.names = F)

# NKI12 (P12)----
met_path <- c("./map_seg_output/nki12_p12")
nki12 <- readVarbinCNA(met_path, remove_Y = TRUE, clean_names = FALSE)
nki12 <- filterCells(nki12, resolution = .8)
nki12 <- findNormalCells(nki12)
# reading metrics
nki12_metrics <- read.table(paste0(met_path, "/metrics/all_stat_metrics.txt"),
                            sep = "\t",
                            header = T) %>% 
  janitor::clean_names() %>% 
  dplyr::rename(sample = "sample_name") %>% 
  mutate(dups_percentage = dups_removed/total_reads) %>% 
  mutate(sample = str_replace_all(sample, '-', '\\.'))

colData(nki12)$timepoint <- ifelse(str_detect(colData(nki12)$sample, "RE"), "recurrence", "primary")
nki12_metadata <- as.data.frame(colData(nki12))
# joining with metrics
nki12_metadata_metrics <- left_join(nki12_metadata, nki12_metrics)
write.table(nki12_metadata_metrics, "./metrics/nki12_p12_metadata.metrics_newnormal.txt", sep = "\t", row.names = F)

nki12 <- nki12[,SummarizedExperiment::colData(nki12)$filtered == "kept"]
nki12@colData %>% as.data.frame() %>% pull(is_normal) %>% table()
nki12 <- nki12[,SummarizedExperiment::colData(nki12)$is_normal == FALSE]
bin_coords2 <- bin_coords %>% dplyr::filter(chrom != 24)
write.table(cbind(bin_coords2, assay(nki12, 'bin_counts')), "./metrics/nki12_p12_filtered_bincounts_newnormal.txt", row.names = F)
